Move files to CLI
