+++ packages/cli/src/commands/hydrogen/dev.ts
@@ -1,7 +1,7 @@
 import path from 'path';
 import fs from 'fs/promises';
 import {outputDebug, outputInfo} from '@shopify/cli-kit/node/output';
-import {fileExists, glob, readFile, writeFile} from '@shopify/cli-kit/node/fs';
+import {fileExists, glob} from '@shopify/cli-kit/node/fs';
 import {renderFatalError} from '@shopify/cli-kit/node/ui';
 import colors from '@shopify/cli-kit/node/colors';
 import {copyPublicFiles} from './build.js';
@@ -21,13 +21,9 @@ import {spawnCodegenProcess} from '../../lib/codegen.js';
 import {getAllEnvironmentVariables} from '../../lib/environment-variables.js';
 import {getConfig} from '../../lib/shopify-config.js';
 import {
-  createMetaobjectDefinition,
-  getMetaobjectDefinitions,
-  updateMetaobjectDefinition,
-} from '../../lib/graphql/admin/metaobject-definitions.js';
-import type {MetaobjectDefinition} from '../../lib/graphql/admin/types-admin-api.js';
-import {SectionSchema} from '../../lib/graphql/admin/types.js';
-import {upsertMetaobject} from '../../lib/graphql/admin/metaobjects.js';
+  getMDForSections,
+  handleSchemaChange,
+} from '../../lib/metaobjects/index.js';
 
 const LOG_REBUILDING = '🧱 Rebuilding...';
 const LOG_REBUILT = '🚀 Rebuilt';
@@ -295,113 +291,4 @@ async function runDev({
       },
     },
   );
-}
-
-const HACK_SESSION = {
-  storeFqdn: 'hydrogen-preview.myshopify.com',
-  token: process.env.HACK_ACCESS_TOKEN as string,
-};
-
-async function handleSchemaChange(
-  file: string,
-  metaobjectDefinitions: Record<string, any>,
-) {
-  console.log('');
-  const {defineSection} = await import('@shopify/hydrogen');
-  const originalFileContent = await readFile(file);
-  const fileContentWithoutImports = originalFileContent
-    .replace(/import\s+[^\s]+\s+from\s+['"][^'"]+['"];?/gims, '')
-    .replace('defineSection', '')
-    .trim();
-
-  // TODO: URI import in Node doesn't seem to support `import` statements
-  const mod = await import(
-    'data:text/javascript;base64,' + btoa(fileContentWithoutImports)
-  );
-
-  const sectionSchema = mod.default as SectionSchema;
-
-  // console.log('new', mod.default);
-  // console.log('old', metaobjectDefinitions[mod.default.type]);
-
-  if (hasMDChanged(sectionSchema, metaobjectDefinitions[sectionSchema.type])) {
-    if (metaobjectDefinitions[sectionSchema.type]) {
-      // Update MD
-      metaobjectDefinitions[sectionSchema.type] =
-        await updateMetaobjectDefinition(
-          HACK_SESSION,
-          sectionSchema,
-          metaobjectDefinitions[sectionSchema.type],
-        );
-    } else {
-      // Create MD
-      metaobjectDefinitions[sectionSchema.type] =
-        await createMetaobjectDefinition(HACK_SESSION, sectionSchema);
-    }
-
-    await upsertMetaobject(HACK_SESSION, sectionSchema);
-  } else {
-    console.log('NO CHANGE FOR', sectionSchema.type);
-  }
-
-  // @ts-expect-error same type coming from different type sources
-  const result = defineSection(sectionSchema);
-  const queryName =
-    sectionSchema.name!.replace(/\s/g, '_').toUpperCase() + '_QUERY';
-  const schemaQuery = mod[queryName] as string | undefined;
-
-  if (result.query !== schemaQuery) {
-    let content = originalFileContent;
-    if (schemaQuery) {
-      // drop the old query
-      content = (content.split(`export const ${queryName}`)[0] ?? '').trim();
-    }
-
-    await writeFile(
-      file,
-      content + `\nexport const ${queryName} = \`${result.query}\`;\n`,
-    );
-  }
-}
-
-async function getMDForSections() {
-  return (await getMetaobjectDefinitions(HACK_SESSION))
-    .filter((metaobject) => metaobject.type.startsWith('section_'))
-    .reduce((acc, item) => {
-      acc[item.type.replace('section_', '')] = item;
-      return acc;
-    }, {} as Record<string, any>);
-}
-
-function hasMDChanged(newMD: SectionSchema, existingMD: MetaobjectDefinition) {
-  if (newMD && !existingMD) return true;
-
-  if (
-    (['name', 'displayNameKey', 'description'] as const).some(
-      (key) =>
-        (newMD[key] || '') !==
-        (existingMD[key] || '').replace('Section | ', ''),
-    ) ||
-    newMD.fields.length !== existingMD.fieldDefinitions.length
-  ) {
-    return true;
-  }
-
-  for (const existingField of existingMD.fieldDefinitions) {
-    const newField = newMD.fields.find(
-      (newField) => newField.key === existingField.key,
-    );
-
-    if (
-      !newField ||
-      !!newField.required !== !!existingField.required ||
-      (['name', 'description'] as const).some(
-        (key) => (newField[key] ?? '') !== (existingField[key] ?? ''),
-      )
-    ) {
-      return true;
-    }
-  }
-
-  return false;
 }
