+++ packages/cli/src/lib/graphql/storefront/metaobjects-fragments.ts
